chore: move auth to lib
