+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,20 +1,5 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!
-    })
-  ],
-  secret: process.env.NEXTAUTH_SECRET,
-  callbacks: {
-    async redirect({ baseUrl }) {
-      return `${baseUrl}/dashboard`;
-    }
-  }
-};
+import { authOptions } from '@/lib/auth';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
